Добавлен GET запрос по id

+++ app/calendars.js
@@ -4,10 +4,24 @@ const auth = require('../middleware/auth');
 
 const router = express.Router();
 
+
+
 router.get('/', auth, async(req, res)=>{
+  const dateNow = new Date();
   try{
-    const calendars = await Calendar.find({$or: [{author: req.user._id}, {shares: req.user._id}]}).populate('author', 'username');
+    const calendars = await Calendar.find({date: {$gte: dateNow},  $or: [{author: req.user._id}, {shares: req.user._id}] })
+                            .populate('author', 'username')
+                            .sort({date: -1});
     res.send(calendars);
+  } catch(err) {
+    res.status(500).send(err);
+  }
+});
+
+router.get('/:id', auth, async(req, res)=>{
+  try{
+    const calendar = await Calendar.findById(req.params.id);
+    res.send(calendar);
   } catch(err) {
     res.status(500).send(err);
   }
@@ -16,9 +30,8 @@ router.get('/', auth, async(req, res)=>{
 router.post('/', auth, async(req, res)=>{
   const calendar = new Calendar(req.body);
   calendar.author = req.user._id;
-
   try {
-    await Calendar.save(calendar);
+    await calendar.save();
     res.send(calendar);
   } catch(err) {
     res.status(500).send(err);
@@ -28,8 +41,13 @@ router.post('/', auth, async(req, res)=>{
 
 router.delete('/:id', auth, async(req, res) => {
   try {
-    await Calendar.findByIdAndDelete(req.params.id);
-    res.send({message: "Successfully deleted"});
+    let calendar = await Calendar.findById(req.params.id);
+    if(req.user._id !== calendar.author._id) {
+      await Calendar.findByIdAndDelete(req.params.id);
+      res.send({message: "Successfully deleted"});
+    } else {
+      res.status(401).send({error: "Access denied!"});
+    }    
   } catch(err) {
     res.status(500).send(err);
   }
@@ -41,7 +59,7 @@ router.put('/:id', auth, async(req, res) => {
     calendar.title = req.body.title;
     calendar.text = req.body.text;
     calendar.date = req.body.date;
-    await Calendar.save(calendar);
+    await calendar.save();
     res.send(calendar);
   } catch(err) {
     res.status(500).send(err);
